refactor: extract tutorial scroll logic and fix strict type errors
