+++ .gitignore
@@ -135,3 +135,4 @@ worktrees/
 .serena/
 build-output.log
 
+\n# Local tool logs\n.bdg/
